Prevent the propagation kernel from deadlocking
It was possible for cos(Cherenkov angle) to be > 1 (due to floating point rounding errors), thus leading to a sampled direction of (nan,nan,nan). This used to be no problem before the implementation of ice anisotropy (photons were being ignored and the effect was minimal). But the anisotropy scattering length correction code could lead the overall scattering length to become NAN, thus trapping the propagation code in an infinite loop. This is only a problem for photon generators setting v/c != 1 (i.e. Geant4).

git-svn-id: http://code.icecube.wisc.edu/svn/projects/clsim/trunk@98377 16731396-06f5-0310-8873-f7f720988828