adhere to OpenCL language standards to fix clsim on Intel OpenCL
The __read_only and __write_only keywords are only valid for image*_t types in OpenCL, whereas I used this for all datatypes in the propagation kernel code. The AMD and nVidia compilers don't really care, but the Intel OpenCL implementation throws a compiler error. This should fix that and make things work with Intel OpenCL.

git-svn-id: http://code.icecube.wisc.edu/svn/projects/clsim/trunk@117555 16731396-06f5-0310-8873-f7f720988828