Emit I3CompressedPhotons if PhotonHistoryEntries == 0.
This reduces the part of I3CLSimModule's memory footprint that is proportional
to deposited energy by a factor of 6. The same applies to the output file
size, if photons are being stored. This could be increased to 10 if we were
willing to store the photon time in single precision (implies short events)
and omit the source particle IDs. Closes #1978.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/clsim/trunk@154598 16731396-06f5-0310-8873-f7f720988828